+++ firmware/utils.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 
 // Absolute time in milliesconds, rough timer
-extern uint64_t abs_time;
+extern uint32_t abs_time;
 
 #define ABS(x) ((x<0)?(-x):x)
